allow exiting process_forever gracefully
hi,
I wanted to use an IRC connection as a subroutine in some code I'm writing, but in all the examples I looked at, whenever the task was done, the program exits via ```sys.exit()``` which is not an option for me. So I added an ```active``` flag to the ```IRC``` class that can be set to false to stop ```process_forever()```. Preliminary tests indicate that it's working as intended :)